third_party/blink/web_tests/platform/generic/fast/filesystem/workers/file-from-file-entry-expected.txt

[Worker] Obtaining File from FileEntry.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/file-from-file-entry.js
PASS [Worker] testFile.name is testFileEntry.name
PASS [Worker] testFile.type is 'text/plain'
PASS [Worker] testFile.size is 0
PASS successfullyParsed is true

TEST COMPLETE
